Reorganized App components
