+++ src/components/Navbar.js
@@ -1,0 +1,11 @@
+import './Navbar.css';
+
+function Navbar() {
+  return (
+    <div className="Nav">
+      <navbar>Hello</navbar>
+    </div>
+  );
+}
+
+export default Navbar;
